357.1. Fetching the Deletion Stream(Section 25: Using React Portals)
In last video we made sure that a user goes to your l like streams to lead and then an idea any time

they attempt to delete a stream.

So now we need to start to refactor the stream delete component.

We are going to turn it into a class based component and we're going to make sure that any time it gets

rendered to the screen it attempts to fetch the stream with the idea of three or whatever is actually

up year.

So we can show the title of it inside the body or content area of this modal.

All right.

So to get started I'm going to open up my stream delete component.

We're going to have to react to this over to a class based component because now we want to make use

of a lifecycle method component in Mt.

So we know when we should attempt to fetch or load up our stream.

So I'm going to delete the stream delete you're at the top and I will replace it with class stream delete

extends re-act component and then I'll open up a curly brace.

Then inside of this mishmash of syntax inside of here right now we need to place our render function.

We're going to put actions right here inside of a helper method.

So I'm going to first find the return keyword right here and I'm going to wrap it with a render method

so I can in place render opening curly brace.

I'll then go to the far side of the return statement and place a closing curly brace like so.

So there's the opening and there's the closing and then technically I really should indent the entire

return statement like so I can also remove the semi-colon at the very end of the class because semi-colons

are not typically required with class statements although technically we can't throw them in there.

Next I'm going to make sure that I wrap actions right here inside of a helper method as well.

So I will say something like render actions and will wrap up actions all indented and then finally we

don't really need to assign this to a variable anymore.

So I'm going to just return parentheses.

Print the C with the re-act fragment like so.

So now down inside the render method last little part of the refactor here when we say actions right

here and pass off the actions variable there is no more actions variable.

I'm going to instead call this render actions and make sure you call it like so we're not trying to

pass a reference off to this function we're trying to call it and pass the result of it to the actions

Propp.

All right let's do a quick Savior and that will refresh our browser and make sure that our refactor

was done successfully.

OK so I can come back over refreshed the page and I do not see any error message.

So if you see any error message right now as usual please read that error message it's going to tell

you exactly what typo you put inside of this class when we refactored it.