358.1. Conditionally Showing Stream Details(Section 25: Using React Portals)
Now that we have fetched our appropriate stream anytime that we come to our streams delete slash ID

route we can show some details about that stream on the motile so that the user knows exactly what stream

they are actually deleting as usual to get some information from our redux store into our component.

We're going to define a map state to prop's function.

So I will flip back over to it stream delete.

I'm going to go down to the bottom and I will define map state to prop's right now what should we call

essentially doing the same exact thing that we went through back on the stream edit components.

So when we call map stage props right here or when this function gets called by re-act redux it's going

to be called with both the state out of our redux store and the second argument of own props and remember

own props right here is the exact same props object that is going to be passed to our component.

The reason we want props is so that we can look at that match prop right there and pull out the idea

of the stream that we're supposed to be showing on this page.

So inside of Matt state the props are going to return an object with the key of stream.

We're going to look at state DOT streams.

Remember state that streams is an object with all of the different streams that we've loaded up.

And if you need a quick reminder that you can always flip back over to your redux step tools click on

that state button right there and then expand streams.

So remember we've got an object right here where the keys are the IDs of each stream.

So now I will reference state data streams and I want to get States streams which specifically the key

of three and we can get that key of three by looking at own props and pulling out the match forams id

property out there from the Ural essentially I'll say square brackets own props match Priam's Id like

the.

All right.

Now I'm not going to forget the next step even though I forgot it like three or four times in this course

so far.

We're going to make sure that we pass maps straight to prop's into connect.

As the first argument.

So map state to prop's like so.

All right so now we can make sure that we attempt to show this stream or some details about the stream

inside of our motile.

Now a quick reminder when this component first gets rendered to the screen it's entirely possible that

we will not have our stream loaded up just yet.

Because remember the component is going to be rendered to the screen one time then our component in

MT lifecycle method will be called.

So the first time this thing gets rendered It's entirely possible that we will not have our stream yet.