358.2. Conditionally Showing Stream Details(Section 25: Using React Portals)
So to handle that there's a couple of different things we can do.

We can either take a similar approach to what we've done in the past where we can do something like

if there is no destock props thought stream then return a div with a text loading.

But this is just a little bit awkward.

The reason for that is that now anytime the user first loads up our application they're going to see

like that div with loading up here and then the modem is going to appear it would be maybe just a little

bit better if we showed the motile right away and maybe we just don't have the text for the stream that

is supposed to be displayed.

So the user will see a modal right away and then like a half second later the title of the stream will

appear inside the content.

I think that would be a little bit less jarring to the user rather than showing them like some loading

text up here and then popping up the modal but realistically we could do it absolutely either way.