358.3. Conditionally Showing Stream Details(Section 25: Using React Portals)
So in order to always show the modal and then just make sure that we update the content field with the

title of the stream I'm going to delete that if statements and I'm going to define another helper method

inside of your called render content so as you might guess render content is going to produce whatever

text it's a post be passed up to the content Propp of modal.

So now inside if you're all say if we do not yet have a stream loaded up then return.

Are you sure you want to get single quotes in here.

Not double.

Are you sure you want to delete this stream.

And then if we have loaded that thing up we can return.

Are you sure you want to delete the stream with title and then we can protect the title inside of your.

To print that out.

We should probably use a E.S. 20:15 template string.

So going to replace those single quotes with back to X and then after that what we say.

Are you sure you want to delete the stream title colon and then we'll do some interpellation here and

we'll say this not prop's stream title.

OK that's a really long line Salumi zoom out so you can see the entire thing.

So again render content is going to say if we do not yet have a stream just print out this kind of like

standon message and then once we actually fetch the stream put the title of the stream into the message

and zoom back in.

And now for content right here rather than just directly assigning a string we can instead place some

curly braces and then make a call to this dot render content like so.

Now last thing I want to point out here is that inside of a render method we currently have the text

stream delete.

And you'll notice that inside of our browser that text is showing up right over here in the background

which is really probably not ideal.

We just don't need to show that text at all.

So I'm going to delete that text of stream delete from the render method.

In addition we've got the wrapping div right here.

We really don't even need the wrapping div at all as well there's no point to that div because nothing

actually shows up inside of it.

Remember our modal uses a re-act portal.

So anything that gets rendered by the modal is going to go directly to the div that we had put into

our indexed HTL file the one with the ID of modal that was the entire point of that portal.

So there's really no point to having this device here at all.

So I going to delete that wrapping as well and I'm left with just returning our modal component.

All right.

I think it's time for a test here.

So going to save this and we'll flip back over it and we'll see how we're doing.

OK.

So back over here my reload kicked in right away.

And I see.

Are you sure you want to delete the stream a title.

I created this stream and so now we could click delete or cancel.

Of course clicking those don't do anything just yet but we'll wired that up in just a second just to

make sure that this works correctly.

I'm going to refresh the page and I still see the title appear right there.

You can also notice that the text can have very suddenly changes like instantaneously because when we

first render the modal to the screen we've got that initial content of.

Are you sure you want to delete the stream and then only after we fetch the thing do we add in the title

at the end.

That's why you see that kind of little blink inside there.

OK so this looks pretty good.

Now the user knows exactly what stream they are attempting to delete.

So unless we have to do is make sure that we wire up the delete and cancel buttons right here.

So let's take care of that in the next video.