359.2. Deleting a Stream(Section 25: Using React Portals)
Now SE1 do very quickly.

Let's take care of making sure that we navigate the user back over to the list of streams after the

stream is deleted.

So I can go back over to my actions index not just file.

So the action Kreator file I'll go down to the bottom where at the delete stream action creator and

we'll just out in history push.

So now after a user deletes the stream we're going to automatically navigate them back to our big list

of streams.

All right so I'll save this and we'll flip back over and do one last quick test here.

Now I don't have any more streams that I can delete so I'm going to create one very quickly.

I'll make a stream with the title of I don't know delete me delete me as the description.

I'll submit it.

Now click on Delete delete again and I get returned back to my big list of streams.

All right.

That's pretty much it.

Not bad.

So we've now got some major Pinnacle's of creating streams listing them editing deleting them all that

kind of good stuff.

I think the last big thing we need to work on is having the ability to actually show a stream to a user

and play some video.

So this is of course going to be a pretty interesting little part.

Let's take a pause right here and we'll start to tackle this part in the next video.