361.1. Switches with React-Router(Section 26: Implementing Streaming Video)
In the last video we got our links working for each of these different streams so I can now click on

a title and see streams show appear.

Now before we start to develop streams show too much I want to show you some really interesting behavior

here is something that's probably going to be pretty unexpected.

I going to go back to my list of streams and then I'm going to click on the Create stream button on

the bottom right hand side and when I do so I see the Create stream component but you'll notice at the

bottom we see something really interesting we see stream show.

So it appears that when we go to stream slash new we see both the Create stream component and streams

show at the same exact time.

So what's going on here.

Well this is kind of one of the downsides to react router and how they choose to implement navigation

with these different route objects.

All right so let's flip back over to our J.S. file.

So inside of apts just remember we have all of our route configurations so essentially we are running

into a scenario where if we go to streams slash new we see both stream create but we also see stream

show.

Now I want to remind you when we first started talking about re-act router Dom we spoke about how re-act

router tries to match these different paths greedily.

In other words if a given path matches a route the route will be shown no matter what re-act router

Dohm doesn't try to just show one route.

It's going to show every single route that matches the given path.

So with that in mind I want you to look at streams slash new right here look at that route.

So streams slash new and and look at the route or the path of streams show down here at the bottom streams

slash wild card are kind of like variable here on the end.

Now up to this point we've been accessing routes of like streams slash five or 20 or three or 100 or

whatever the I.D. was.

But remember the Colan right here is essentially a variable.

It does not only match variables that are numbers it will match anything in that part of the or else.

So in this case this route right here with streams show when we go to stream slashed new ID right here

is essentially taking the text new because it thinks New Right here is like a variable or a idea of

sorts.

And so that's why you're seeing both stream create and streams show.