361.2. Switches with React-Router(Section 26: Implementing Streaming Video)
So we need to figure out some way to only show either streams new or stream slash colon ID in order

to do so we're going to import another helper component from re-act Rotterdam.

So up at the top I'm going to find my re-act writer Dom in port and I'm going to also import something

called switch.

So back down at my route definition I'm going to wrap all of these routes inside of that switch thing

that we just imported.

So I'll say switch opening tag and then at the bottom of all of our routes open and another switch like

so so switch right here is going to look at all these different routes and it's only going to show one

of these given routes for any path that we go to.

So the first route inside of your that gets Matt matched by a given path is going to be shown and nothing

else.

So now whenever we go to stream slash knew this route will be shown and re-act Rotterdam is going to

say OK I found a route right here to show I'm not going to show any other potential route.

So even though technically stream's slash ID colon ID would match stream slash knew it will not be shown

because it's now wrapped inside of the switch.

So now we can save this flip back over and when we go to stream slash knew we will no longer see streams

show at the bottom as we would probably expect.

OK.

So this is a problem that I really expect you're going to run into when you start working on your own

project.

Remember any time that it feels like re-act Rotterdam is showing a component that it shouldn't show.

Chances are you've got one of these little query parameters on here and that's probably what is throwing

off your route matching.

And you can very easily solve this by just adding in a switch to wrap all of your different routes.

OK.

Now that we've got this put together let's take a break when we come back the next video.

We'll continue working on our streams show component.